java: fix warnings from intellij

+++ jackson-jq-core/src/main/java/net/thisptr/jackson/jq/v2/core/internal/tree/matcher/SlotResolver.java
@@ -34,6 +34,6 @@ public final class SlotResolver {
 			throw new IllegalStateException("No slot allocated for pattern variable $" + name);
 		if (!claimed.add(name))
 			return -1;
-		return slot.intValue();
+		return slot;
 	}
 }
